Calculate a score for every city and display in cities panel
Score is calculated for each city by multiplying the city’s att rank by
attWeight for every checked attribute and then taking the mean of all
those values. This currently displayed in the cities panel. Need to
restructure and actually figure out how we want to deal with weight and
scoring according to rank.